# docs/swagger.yaml
definitions:
  controller.albumRequest:
    properties:
      length:
        type: integer
      name:
        type: string
    required:
    - length
    - name
    type: object
  controller.authRequest:
    properties:
      email:
        type: string
      password:
        minLength: 8
        type: string
    required:
    - email
    - password
    type: object
  controller.createUserRequest:
    properties:
      email:
        type: string
      name:
        type: string
      password:
        minLength: 8
        type: string
    required:
    - email
    - name
    - password
    type: object
  controller.updateUserRequest:
    properties:
      email:
        type: string
      name:
        type: string
    required:
    - email
    - name
    type: object
  domain.Album:
    properties:
      created_at:
        type: string
      id:
        type: string
      length:
        type: integer
      name:
        type: string
      updated_at:
        type: string
    type: object
  domain.AuthToken:
    properties:
      token:
        type: string
    type: object
  domain.User:
    properties:
      created_at:
        type: string
      email:
        type: string
      id:
        type: string
      name:
        type: string
      password:
        type: string
      updated_at:
        type: string
    type: object
  rest.Message:
    properties:
      message:
        type: string
      status:
        type: integer
    type: object
host: localhost:8000
info:
  contact:
    email: xorycx@gmail.com
    name: Cyro Dubeux
    url: https://github.com/cyruzin/hexagony/issues/new
  description: Clean architecture example in Golang.
  license:
    name: MIT
    url: https://github.com/cyruzin/hexagony/blob/master/LICENSE
  title: Hexagony API
  version: "1.0"
paths:
  /album:
    get:
      consumes:
      - application/json
      description: lists all albums
      parameters:
      - default: Bearer <Add access token here>
        description: Insert your access token
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/domain.Album'
            type: array
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/rest.Message'
      summary: List of albums
      tags:
      - album
    post:
      consumes:
      - application/json
      description: add a new album
      parameters:
      - default: Bearer <Add access token here>
        description: Insert your access token
        in: header
        name: Authorization
        required: true
        type: string
      - description: add a new album
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/controller.albumRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/rest.Message'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/rest.Message'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/rest.Message'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/rest.Message'
      summary: Add an album
      tags:
      - album
  /album/{uuid}:
    delete:
      consumes:
      - application/json
      description: delete an album by uuid
      parameters:
      - default: Bearer <Add access token here>
        description: Insert your access token
        in: header
        name: Authorization
        required: true
        type: string
      - description: album uuid
        in: path
        name: uuid
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/rest.Message'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/rest.Message'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/rest.Message'
      summary: Delete an album
      tags:
      - album
    get:
      consumes:
      - application/json
      description: lists an album by uuid
      parameters:
      - default: Bearer <Add access token here>
        description: Insert your access token
        in: header
        name: Authorization
        required: true
        type: string
      - description: album uuid
        in: path
        name: uuid
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.Album'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/rest.Message'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/rest.Message'
      summary: List an album
      tags:
      - album
    put:
      consumes:
      - application/json
      description: update an album by uuid
      parameters:
      - default: Bearer <Add access token here>
        description: Insert your access token
        in: header
        name: Authorization
        required: true
        type: string
      - description: album uuid
        in: path
        name: uuid
        required: true
        type: string
      - description: update an album by uuid
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/controller.albumRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/rest.Message'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/rest.Message'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/rest.Message'
      summary: Update an album
      tags:
      - album
  /auth:
    post:
      consumes:
      - application/json
      description: authenticate a user and returns a JWT token
      parameters:
      - description: authenticates the user
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/controller.authRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.AuthToken'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/rest.Message'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/rest.Message'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/rest.Message'
      summary: Authenticate a user
      tags:
      - auth
  /user:
    get:
      consumes:
      - application/json
      description: lists all users
      parameters:
      - default: Bearer <Add access token here>
        description: Insert your access token
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/domain.User'
            type: array
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/rest.Message'
      summary: List of users
      tags:
      - user
    post:
      consumes:
      - application/json
      description: add a new user
      parameters:
      - default: Bearer <Add access token here>
        description: Insert your access token
        in: header
        name: Authorization
        required: true
        type: string
      - description: add a new user
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/controller.createUserRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/rest.Message'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/rest.Message'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/rest.Message'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/rest.Message'
      summary: Add an user
      tags:
      - user
  /user/{uuid}:
    delete:
      consumes:
      - application/json
      description: delete an user by uuid
      parameters:
      - default: Bearer <Add access token here>
        description: Insert your access token
        in: header
        name: Authorization
        required: true
        type: string
      - description: user uuid
        in: path
        name: uuid
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/rest.Message'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/rest.Message'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/rest.Message'
      summary: Delete an user
      tags:
      - user
    get:
      consumes:
      - application/json
      description: lists an user by uuid
      parameters:
      - default: Bearer <Add access token here>
        description: Insert your access token
        in: header
        name: Authorization
        required: true
        type: string
      - description: user uuid
        in: path
        name: uuid
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.User'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/rest.Message'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/rest.Message'
      summary: List an user
      tags:
      - user
    put:
      consumes:
      - application/json
      description: update an user by uuid
      parameters:
      - default: Bearer <Add access token here>
        description: Insert your access token
        in: header
        name: Authorization
        required: true
        type: string
      - description: user uuid
        in: path
        name: uuid
        required: true
        type: string
      - description: update an user by uuid
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/controller.updateUserRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/rest.Message'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/rest.Message'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/rest.Message'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/rest.Message'
      summary: Update an user
      tags:
      - user
swagger: "2.0"